Users/rajes/OneDrive/Desktop/Auto post/files (12)/requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
aiohttp==3.9.3
